refactor(layering): evacuate odbc/scylladb/imap/git providers into their connector crates
Move each connector's provider code out of the data_components monolith into its
already-wired connector-* crate, so the driver deps leave data_components' graph:

- odbc:     data_components/src/odbc.rs -> connector-odbc (odbc-api stays shared in
            db_connection_pool; no restricted_deps entry needed).
- scylladb: data_components/src/scylladb/ -> connector-scylladb; move the scylla
            provider dep + drop the data_components 'scylladb' feature; runtime keeps
            only db_connection_pool/scylladb (the pool). scylla stays a shared pool
            driver, like odbc-api.
- imap:     data_components/src/imap/ -> connector-imap; move charset/imap/mailparse
            drivers; add restricted_deps entries.
- git:      data_components/src/git.rs -> connector-git; move git2/url/arrow/regex/
            dashmap/tempfile deps; add git2 restricted_deps entry.

Rewrites shared-infra refs (crate::{arrow,federation,function_support,key_filter,
rate_limit,Read}) to data_components::*, fixes the now-dangling runtime imap/scylladb
feature pass-throughs, and removes runtime's orphaned scylla dep.

The scylladb/git provider modules are re-exported as 'pub mod' (as they were in
data_components) so clippy's avoid-breaking-exported-api treats their public API items
as exported -- otherwise moving them into a private module surfaces latent dead_code /
enum_variant_names on API that was previously exempt. connector-scylladb gains the same
#![allow(missing_errors_doc)] that data_components and connector-git already carry for
this moved provider code. Removed a genuinely-dead odbc Error import/variants.

Layering guard green (131 crates, no upward edges). No behavior change.
