+++ crates/data-connectors/connector-git/Cargo.toml
@@ -7,11 +7,17 @@ license-file.workspace = true
 description = "Git data connector for Spice.ai runtime"
 
 [dependencies]
+arrow.workspace = true
 async-trait.workspace = true
+dashmap = "6.1"
 data_components = { path = "../../data_components" }
 datafusion.workspace = true
+git2.workspace = true
 globset.workspace = true
 opentelemetry.workspace = true
+regex.workspace = true
+tempfile.workspace = true
+url.workspace = true
 reqwest.workspace = true
 runtime = { path = "../../runtime" }
 runtime-api-types = { path = "../../runtime-api-types" }
