+++ crates/data-connectors/connector-imap/Cargo.toml
@@ -7,10 +7,13 @@ license-file.workspace = true
 description = "IMAP data connector for Spice.ai runtime"
 
 [dependencies]
+arrow.workspace = true
 async-trait.workspace = true
-data_components = { path = "../../data_components", features = ["imap"] }
+charset.workspace = true
+data_components = { path = "../../data_components" }
 datafusion.workspace = true
 imap.workspace = true
+mailparse.workspace = true
 linkme.workspace = true
 paste.workspace = true
 regex.workspace = true
