+++ crates/data_components/Cargo.toml
@@ -34,7 +34,6 @@ bb8 = { workspace = true, optional = true }
 bb8-oracle = { version = "0.3", features = ["chrono"], optional = true }
 bigdecimal = { workspace = true, optional = true }
 bytes.workspace = true
-charset = { workspace = true, optional = true }
 chrono.workspace = true
 connector-postgres-common = { path = "../data-connectors/connector-postgres-common", optional = true }
 dashmap = "6.1"
@@ -55,7 +54,6 @@ duckdb = { workspace = true, features = [
 ], optional = true }
 flight_client = { path = "../flight_client" }
 futures.workspace = true
-git2 = { workspace = true }
 globset.workspace = true
 graph-rs-sdk = { workspace = true, optional = true }
 graphql-parser.workspace = true
@@ -66,12 +64,10 @@ iceberg.workspace = true
 parquet = { workspace = true }
 iceberg-catalog-rest.workspace = true
 iceberg-datafusion.workspace = true
-imap = { workspace = true, optional = true }
 indexmap.workspace = true
 logos.workspace = true
 mysql_async = { workspace = true, optional = true }
 mongodb = { workspace = true, optional = true }
-mailparse = { workspace = true, optional = true }
 num_cpus = "1.16"
 object_store = { workspace = true }
 opendal.workspace = true
@@ -93,7 +89,6 @@ rusqlite = { workspace = true, optional = true }
 elasticsearch = { path = "../elasticsearch", optional = true }
 s3_vectors = { path = "../s3_vectors", optional = true }
 s3_vectors_metadata_filter = { path = "../s3_vectors_metadata_filter", optional = true }
-scylla = { workspace = true, optional = true }
 secrecy.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -149,7 +144,6 @@ duckdb = [
     "datafusion-table-providers/duckdb-federation",
 ]
 flightsql = []
-imap = ["dep:charset", "dep:imap", "dep:mailparse"]
 kafka = ["dep:rdkafka"]
 mongodb = ["dep:mongodb", "datafusion-table-providers/mongodb"]
 mssql = ["dep:tiberius", "dep:tokio-util", "dep:bb8"]
@@ -158,7 +152,6 @@ mysql = [
     "datafusion-table-providers/mysql",
     "datafusion-table-providers/mysql-federation",
 ]
-odbc = ["federation", "db_connection_pool/odbc"]
 oracle = ["dep:oracle", "dep:bb8", "dep:bb8-oracle", "dep:bigdecimal"]
 postgres = [
     "dep:tokio-postgres",
@@ -172,7 +165,6 @@ postgres = [
     "datafusion-table-providers/postgres-federation",
 ]
 s3_vectors = ["dep:s3_vectors", "dep:s3_vectors_metadata_filter"]
-scylladb = ["dep:scylla", "datafusion-table-providers/federation"]
 sharepoint = ["dep:graph-rs-sdk"]
 # Enables graph-rs-sdk's `test-util` feature so `GraphClient::use_test_endpoint`
 # accepts non-Microsoft hosts (e.g. localhost mocks). Intended for tests only.
