+++ crates/data_components/src/lib.rs
@@ -152,8 +152,6 @@ pub mod mssql;
 pub mod mysql;
 #[cfg(feature = "mysql")]
 pub mod mysql_replication;
-#[cfg(feature = "odbc")]
-pub mod odbc;
 #[cfg(feature = "oracle")]
 pub mod oracle;
 #[cfg(feature = "postgres")]
@@ -170,8 +168,6 @@ pub mod schema_discovery;
 /// `datafusion-table-providers` fork so providers defined there (`MongoDB`) can
 /// reuse it; re-exported here for the in-repo connectors (`DynamoDB`, Debezium).
 pub use datafusion_table_providers::schema_projection;
-#[cfg(feature = "scylladb")]
-pub mod scylladb;
 pub mod sql_expr;
 
 #[cfg(feature = "sharepoint")]
@@ -187,7 +183,6 @@ pub mod sqlite;
 pub mod turso;
 pub mod unity_catalog;
 
-pub mod git;
 pub mod github;
 pub mod key_filter;
 pub mod pk_filter_expr;
@@ -197,8 +192,6 @@ pub mod cdc;
 pub mod delete;
 pub mod graphql;
 pub mod http;
-#[cfg(feature = "imap")]
-pub mod imap;
 pub mod index_maintenance;
 pub mod object;
 pub mod poly;
